requirements.txt
PyQt5==5.15.1
numpy==1.19.4
Pillow==8.0.1
scikit-learn==0.23.2
